+++ jscomp/runtime/js.pre.mli
@@ -232,6 +232,9 @@ module WeakMap = Js_weakmap
 module Iterator = Js_iterator
 (** Bindings to functions on [Iterator] *)
 
+module FormData = Js_formData
+(** Bindings to FormData *)
+
 (**/**)
 
 module OO = Js_OO